pyproject.toml
[tool.poetry]
name = "tochka_test"
version = "0.1.0"
description = "flask test app for tochka bank"
authors = ["Anisimov Danil <frosthamster@gmail.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
sqlalchemy = "^1.3"
marshmallow = "^3.2"
flask-uuid = "^0.2.0"
simplejson = "^3.16"
flask-sqlalchemy = "^2.4"
flask-migrate = "^2.5"
marshmallow-sqlalchemy = "^0.19.0"
redis = "^3.3"
huey = "^2.1"
python-dotenv = "^0.10.3"
gunicorn = "^20.0"
psycopg2-binary = "^2.8"
invoke = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-xdist = "^1.30"
pytest-cov = "^2.8"
pytest-deadfixtures = "^2.1"
pytest-mock = "^1.11"
pytest-flask = "^0.15.0"
pytest-flask-sqlalchemy = "^1.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
unify = "^0.5.0"
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
